refactor #189: 엔티티 수정 : BaseTimeEntity->BaseEntity / 삭제시 권한 확인 추가

+++ src/main/java/com/irum/come2us/domain/store/application/service/StoreService.java
@@ -61,6 +61,7 @@ public class StoreService {
 
     public void withdrawStore(UUID storeId) {
         Store store = getStoreById(storeId);
+        memberUtil.assertMemberResourceAccess(store.getMember());
         storeRepository.delete(store);
     }
